Update LED Display.py
Corrected issue where the line drawing function would skip the index y=64 (top row of the lower row of panels)

+++ LED Display.py
@@ -39,8 +39,8 @@ class LEDdisplay(SampleBase):
             distance = int(math.sqrt( (x_distance)**2 + (y_distance)**2 ))
             
             for i in range(0,distance*2):
-                Px = a[0] + (x_distance/2)*(i/distance)
-                Py = a[1] + (y_distance/2)*(i/distance)
+                Px = int(a[0] + (x_distance/2)*(i/distance))
+                Py = int(a[1] + (y_distance/2)*(i/distance))
 
                 print(Py)
